+++ resources/views/bootstrap-5/components/forms/inputs/password.blade.php
@@ -1,0 +1,10 @@
+<input
+    name="{{ $name }}"
+    type="password"
+    id="{{ $id }}"
+    @if($hasErrors)aria-describedby="validation-{{ $name }}-feedback"@endif
+    {{ $attributes->class([
+        'form-control',
+        'is-invalid' => $hasErrors,
+    ]) }}
+/>