+++ resources/views/bootstrap-5/components/forms/inputs/textarea.blade.php
@@ -1,0 +1,10 @@
+<textarea
+    name="{{ $name }}"
+    id="{{ $id }}"
+    rows="{{ $rows }}"
+    @if($hasErrors)aria-describedby="validation-{{ $name }}-feedback"@endif
+    {{ $attributes->class([
+        'form-control',
+        'is-invalid' => $hasErrors,
+    ]) }}
+>{{ old($name, $slot) }}</textarea>