+++ src/Components/Forms/Error.php
@@ -5,13 +5,14 @@ declare(strict_types=1);
 namespace BladeUIKitBootstrap\Components\Forms;
 
 use BladeUIKitBootstrap\Concerns\CanHaveErrors;
+use BladeUIKitBootstrap\Concerns\HasBootstrapVersion;
 use BladeUIKit\Components\BladeComponent;
-use Illuminate\Support\ViewErrorBag;
 use Illuminate\View\View;
 
 class Error extends BladeComponent
 {
     use CanHaveErrors;
+    use HasBootstrapVersion;
 
     public function __construct(string $field, string $bag = 'default')
     {
@@ -20,6 +21,6 @@ class Error extends BladeComponent
 
     public function render(): View
     {
-        return view('blade-ui-kit-bootstrap::components.forms.error');
+        return view($this->viewPath('components.forms.error'));
     }
 }