+++ src/Components/Forms/Inputs/Date.php
@@ -7,17 +7,17 @@ namespace BladeUIKitBootstrap\Components\Forms\Inputs;
 use BladeUIKitBootstrap\Concerns\HasBootstrapVersion;
 use Illuminate\Contracts\View\View;
 
-class Email extends Input
+class Date extends Input
 {
     use HasBootstrapVersion;
 
     public function __construct(string $name, string $id = null, ?string $value = '', ?string $errorBag = null)
     {
-        parent::__construct($name, $id, 'email', $value, $errorBag);
+        parent::__construct($name, $id, 'date', $value, $errorBag);
     }
 
     public function render(): View
     {
-        return view($this->viewPath('components.forms.inputs.email'));
+        return view($this->viewPath('components.forms.inputs.date'));
     }
 }