+++ src/Components/Forms/Inputs/Input.php
@@ -5,12 +5,14 @@ declare(strict_types=1);
 namespace BladeUIKitBootstrap\Components\Forms\Inputs;
 
 use BladeUIKitBootstrap\Concerns\CanHaveErrors;
+use BladeUIKitBootstrap\Concerns\HasBootstrapVersion;
 use BladeUIKit\Components\Forms\Inputs\Input as BukInput;
 use Illuminate\Contracts\View\View;
 
 class Input extends BukInput
 {
     use CanHaveErrors;
+    use HasBootstrapVersion;
 
     public function __construct(string $name, string $id = null, string $type = 'text', ?string $value = '', ?string $errorBag = null)
     {
@@ -21,6 +23,6 @@ class Input extends BukInput
 
     public function render(): View
     {
-        return view('blade-ui-kit-bootstrap::components.forms.inputs.input');
+        return view($this->viewPath('components.forms.inputs.input'));
     }
 }